requirements.txt
streamlit
pandas
numpy
matplotlib
pyyaml
requests
pillow
